temp_arquivos_antigos_excluir_10dez2025/scripts_antigos/requirements.txt
# App Leonardo - Bot de Trading
# Dependências do Projeto

# Core trading
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0

# Django Dashboard
Django>=4.2.0
django-cors-headers>=4.0.0

# Utilities
python-dotenv>=1.0.0
PyYAML>=6.0.0
requests>=2.31.0

# Visualização (opcional para dashboard terminal)
rich>=13.0.0

# Análise técnica
ta-lib>=0.4.28  # Opcional - pode usar implementação nativa
